+++ vacefron/__init__.py
@@ -1,7 +1,7 @@
-from .classes import Image, RankCard
+from .classes import Image, RankCard, CrewMateColors
 from .client import Client
 from .errors import BadRequest, HTTPException, InternalServerError, NotFound
 
-__license__ = 'MIT'
-__author__ = 'Soheab_'
-__version__ = '1.2.2'
+__license__ = "MIT"
+__author__ = "Soheab_"
+__version__ = "1.3.0"